pyproject.toml
[project]
name = "langchain-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.3",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.16",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
    "tavily-python>=0.7.17",
]
